Refactored the code

+++ app.js
@@ -17,8 +17,12 @@ const posts = [];
 
 const app = express();
 
-var post = mongoose.createConnection("mongodb+srv://Bharath_xD:Saibharat%40123@cluster0.cgaoktp.mongodb.net/blogDB?retryWrites=true&w=majority");
-var user = mongoose.createConnection("mongodb+srv://Bharath_xD:Saibharat%40123@cluster0.cgaoktp.mongodb.net/userDB?retryWrites=true&w=majority");
+var post = mongoose.createConnection(
+  "mongodb+srv://Bharath_xD:Saibharat%40123@cluster0.cgaoktp.mongodb.net/blogDB?retryWrites=true&w=majority"
+);
+var user = mongoose.createConnection(
+  "mongodb+srv://Bharath_xD:Saibharat%40123@cluster0.cgaoktp.mongodb.net/userDB?retryWrites=true&w=majority"
+);
 
 app.set("view engine", "ejs");
 app.use(flash());
@@ -37,36 +41,33 @@ app.use(passport.session());
 passport.use(
   new GoogleStrategy(
     {
-      clientID: "160599315944-c2b9g24bgp8mka1putls852rgivfm8jc.apps.googleusercontent.com",
-      clientSecret:"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",
-      callbackURL: "https://blogger-by-bharath.herokuapp.com/auth/google/compose",
+      clientID:
+        "160599315944-c2b9g24bgp8mka1putls852rgivfm8jc.apps.googleusercontent.com",
+      clientSecret: "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",
+      callbackURL:
+        "https://blogger-by-bharath.herokuapp.com/auth/google/compose",
     },
-    function (accessToken, refreshToken, profile, cb) {
+    (accessToken, refreshToken, profile, cb) => {
       User.findOrCreate(
         { googleId: profile.id, username: profile.displayName },
-        function (err, user) {
+        (err, user) => {
           return cb(err, user);
         }
       );
     }
   )
 );
-app.use(function (req, res, next) {
-  res.locals.error = req.flash("error");
-  res.locals.success = req.flash("success");
-  next();
-});
 
-app.use(function (req, res, next) {
-  if (req.isAuthenticated()) {
-    var username = req.user.username;
-  } else {
-    var username = "";
-  }
-  res.locals.username = username;
-  res.locals.signinStatus = req.isAuthenticated();
-  next();
-});
+const headerStatus = require("./routes/HeaderStatus.js");
+// Updates the changes in the Header
+const authenticate = require("./routes/authenticate.js");
+// Containes Google Authentication routes
+const useFlash = require("./routes/flash.js");
+// Containes the flash messages for login and report routes
+
+app.use(authenticate);
+app.use(headerStatus);
+app.use(useFlash);
 
 // Mongoose Schema for posts
 
@@ -93,54 +94,42 @@ const User = user.model("User", userSchema);
 
 passport.use(User.createStrategy());
 
-passport.serializeUser(function (user, done) {
+passport.serializeUser((user, done) => {
   done(null, user.id);
 });
 
-passport.deserializeUser(function (id, done) {
-  User.findById(id, function (err, user) {
+passport.deserializeUser((id, done) => {
+  User.findById(id, (err, user) => {
     done(err, user);
   });
 });
 
-app.get(
-  "/auth/google",
-  passport.authenticate("google", { scope: ["profile"] })
-);
-app.get(
-  "/auth/google/compose",
-  passport.authenticate("google", { failureRedirect: "/login" }),
-  function (req, res) {
-    res.redirect("/compose");
-  }
-);
-
-app.get("/", function (req, res) {
-  Post.find({}, function (err, foundItems) {
+app.get("/", (req, res) => {
+  Post.find({}, (err, foundItems) => {
     res.render("home", {
       posts: foundItems,
     });
   });
 });
 
-app.get("/logout", function (req, res) {
-  req.logout(function (err) {
-    if (err) {
-      return next(err);
-    } else {
-      setTimeout(function (err) {
-        if (!err) {
-          res.redirect("/");
-        }
-        {
-          console.log(err);
-        }
-      }, 0);
-    }
-  });
+app.get("/about", (req, res) => {
+  res.locals.signinStatus = req.isAuthenticated();
+  res.render("about");
 });
 
-app.get("/login", function (req, res) {
+app.get("/contact", (req, res) => {
+  res.render("contact");
+});
+
+app.get("/register", (req, res) => {
+  if (req.isAuthenticated()) {
+    res.redirect("/");
+  } else {
+    res.render("register");
+  }
+});
+
+app.get("/login", (req, res) => {
   if (req.isAuthenticated()) {
     res.redirect("/");
   } else {
@@ -155,34 +144,33 @@ app.get("/login", function (req, res) {
   }
 });
 
-app.get("/register", function (req, res) {
-  if (req.isAuthenticated()) {
-    res.redirect("/");
-  } else {
-    res.render("register");
-  }
+app.get("/logout", (req, res) => {
+  req.logout((err) => {
+    if (err) {
+      return next(err);
+    } else {
+      setTimeout((err) => {
+        if (!err) {
+          res.redirect("/");
+        }
+        {
+          console.log(err);
+        }
+      }, 0);
+    }
+  });
 });
 
-app.get("/about", function (req, res) {
-  res.locals.signinStatus = req.isAuthenticated();
-  res.render("about");
-});
-
-app.get("/contact", function (req, res) {
-  res.render("contact");
-});
-
-app.get("/compose", function (req, res) {
+app.get("/compose", (req, res) => {
   if (req.isAuthenticated()) {
     res.render("compose");
   } else {
     res.render("login");
   }
 });
-
-app.get("/posts/:name", function (req, res) {
+app.get("/posts/:name", (req, res) => {
   const requestedPostId = req.params.name;
-  Post.findById({ _id: requestedPostId }, function (err, foundPost) {
+  Post.findById({ _id: requestedPostId }, (err, foundPost) => {
     if (!err) {
       res.render("post", {
         postTitle: foundPost.title,
@@ -190,12 +178,12 @@ app.get("/posts/:name", function (req, res) {
         postAuthor: foundPost.author,
       });
     } else {
-      console.log("err at line 193");
+      console.log("err");
     }
   });
 });
 
-app.get("/report", function (req, res) {
+app.get("/report", (req, res) => {
   if (req.isAuthenticated()) {
     var username = req.user.username;
   }
@@ -205,23 +193,16 @@ app.get("/report", function (req, res) {
   });
 });
 
-app.get("/failure", function (req, res) {
-  if (!req.user) {
-    req.flash("success", "Username or password is incorrect.");
-    res.redirect("/login");
-  }
-});
-
-app.post("/register", function (req, res) {
+app.post("/register", (req, res) => {
   User.register(
     { username: req.body.username },
     req.body.password,
-    function (err, user) {
+    (err, user) => {
       if (err) {
         req.flash("error", err.message);
         res.redirect("/register");
       } else {
-        passport.authenticate("local")(req, res, function (err) {
+        passport.authenticate("local")(req, res, (err) => {
           if (!err) {
             res.redirect("/compose");
           } else {
@@ -233,21 +214,21 @@ app.post("/register", function (req, res) {
   );
 });
 
-app.post("/login", function (req, res) {
+app.post("/login", (req, res) => {
   const user = new User({
     username: req.body.username,
     password: req.body.password,
   });
-  req.login(user, function (err) {
+  req.login(user, (err) => {
     if (err) {
       console.log(err);
     } else {
       passport.authenticate("local", {
-        successFlash: "Hey, Welcome back",
+        successFlash: "Welcome!",
         successRedirect: "/compose",
         failureFlash: true,
         failureRedirect: "/login",
-      })(req, res, function (err) {
+      })(req, res, (err) => {
         if (!err) {
           res.redirect("/compose");
         } else {
@@ -258,31 +239,32 @@ app.post("/login", function (req, res) {
   });
 });
 
-app.post("/compose", function (req, res) {
+app.post("/compose", (req, res) => {
   const post = new Post({
     title: req.body.inputTitle,
     author: req.user.username,
     content: req.body.textAreaPost,
   });
 
-  post.save(function (err) {
+  post.save((err) => {
     if (!err) {
       res.redirect("/");
     }
   });
 });
 
-app.post("/report", function (req, res) {
+app.post("/report", (req, res) => {
   const ra = req.body.reportAuthor;
-  Post.findOneAndDelete(
-    { author: req.body.reportAuthor },
-    function (err, post) {
-      req.flash('success', "We have recieved your report :D ");
-      res.redirect("report");
-    }
-  );
+  Post.findOneAndDelete({ author: req.body.reportAuthor }, (err, pot) => {
+    req.flash("success", "We have recieved your report :D ");
+    res.redirect("report");
+  });
 });
 
-app.listen(process.env.PORT || 3000, function(){
-  console.log("Express server listening on port %d in %s mode", this.address().port, app.settings.env);
+app.listen(process.env.PORT || 3000, function () {
+  console.log(
+    "Express server listening on port %d in %s mode",
+    this.address().port,
+    app.settings.env
+  );
 });
